Fix TickTick tag sortOrder overflow: Int -> Long
The TickTick API returns sortOrder values like -1099513921536 which exceed
Int.MAX_VALUE. Change ApiTickTickTag.sortOrder from Int? to Long?, consistent
with all other sortOrder fields across the data layer.

+++ mobile/app/src/main/java/dev/arunkumar/jarvis/data/ticktick/ApiTickTickTag.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class ApiTickTickTag(
   val name: String,
   val label: String? = null,
-  val sortOrder: Int? = null,
+  val sortOrder: Long? = null,
   val sortType: String? = null,
   val color: String? = null,
   val parent: String? = null,
